+++ packages/composites/defi-pulse/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/defi-pulse-adapter
 
+## 1.2.5
+
+### Patch Changes
+
+- Updated dependencies [8161e1e18]
+  - @chainlink/ea-bootstrap@2.15.0
+  - @chainlink/token-allocation-adapter@1.5.5
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 1.2.4
 
 ### Patch Changes