+++ packages/composites/por-indexer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/por-indexer-adapter
 
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies [8161e1e18]
+  - @chainlink/ea-bootstrap@2.15.0
+  - @chainlink/proof-of-reserves-adapter@1.12.2
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 1.2.5
 
 ### Patch Changes