+++ packages/composites/the-graph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/the-graph-adapter
 
+## 1.2.5
+
+### Patch Changes
+
+- Updated dependencies [8161e1e18]
+  - @chainlink/ea-bootstrap@2.15.0
+  - @chainlink/ea-reference-data-reader@1.3.5
+  - @chainlink/ea-test-helpers@1.4.0
+  - @chainlink/graphql-adapter@1.2.5
+
 ## 1.2.4
 
 ### Patch Changes