+++ packages/core/reference-data-reader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/ea-reference-data-reader
 
+## 1.3.5
+
+### Patch Changes
+
+- Updated dependencies [8161e1e18]
+  - @chainlink/ea-bootstrap@2.15.0
+
 ## 1.3.4
 
 ### Patch Changes