+++ packages/sources/ethgasstation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/ethgasstation-adapter
 
+## 1.4.5
+
+### Patch Changes
+
+- Updated dependencies [8161e1e18]
+  - @chainlink/ea-bootstrap@2.15.0
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 1.4.4
 
 ### Patch Changes