+++ packages/sources/synthetix-debt-pool/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/synthetix-debt-pool-adapter
 
+## 3.2.5
+
+### Patch Changes
+
+- Updated dependencies [8161e1e18]
+  - @chainlink/ea-bootstrap@2.15.0
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 3.2.4
 
 ### Patch Changes